Create user move and single round game functions
Create a function that will promt the user for their move/choice (Rock,
Paper, or Scissors) and return a string with only the first letter
capitalized. Create a function that will take the computer's choice
and the user's choice to play a round of Rock, Paper, Scissors. Using
else if statements, account for all possible user and computer selection
combinations. If the user provides an input other than the three moves
allowed, return a message letting the user know that is not an option.