Create a function to play a game of best of 5
Create a function that uses a for loop and calls the playRound function
inside the for loop to play a game of best of 5. Use if...else
statements to add a point to the winning side's score or replay the
round if there is a tie. Break the loop and declare a winner if either
side reaches 3 wins. Remove the remaining pseudocode and leave comments
to identify all of the declared functions.